Move SharedServicesContainer, refs 2275
